+++ data/authors.json
@@ -6,14 +6,14 @@
     "items": [
       {
         "type": "Book",
-        "id": 753,
+        "id": 644,
         "publish_date": "2020-01-01",
         "publisher": "Petro Loltolo",
         "cover_state": "new"
       },
       {
         "type": "Movie",
-        "id": 516,
+        "id": 152,
         "publish_date": "2020-01-01",
         "silent": true
       }
@@ -26,9 +26,37 @@
     "items": [
       {
         "type": "Game",
-        "id": 822,
+        "id": 502,
         "publish_date": "2023-01-01",
         "last_played_at": "2022-12-02",
+        "multiplayer": true
+      }
+    ]
+  },
+  {
+    "id": 231,
+    "first_name": "ZikaZaki",
+    "last_name": "Ali",
+    "items": [
+      {
+        "type": "Game",
+        "id": 826,
+        "publish_date": "2023-12-02",
+        "last_played_at": "2019-12-02",
+        "multiplayer": false
+      }
+    ]
+  },
+  {
+    "id": 670,
+    "first_name": "Zack",
+    "last_name": "Ali",
+    "items": [
+      {
+        "type": "Game",
+        "id": 158,
+        "publish_date": "2024-12-02",
+        "last_played_at": "2019-02-02",
         "multiplayer": true
       }
     ]
